app/python/pyproject.toml
[project]
name = "scrcpy-ai"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "httpx>=0.27.0",
    "websockets>=12.0",
    "openai>=1.50.0",
    "open-clip-torch>=2.26.0",
    "Pillow>=10.0.0",
    "numpy>=1.26.0",
    "chromadb>=0.5.0",
    "imagehash>=4.3.0",
]

[project.scripts]
scrcpy-ai = "scrcpy_ai.main:main"
